+++ CHANGELOG.md
@@ -10,6 +10,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Added
 
 -   Checkbox in global Settings for disabling camera reset, when new map is loaded #685
+
+### Changed
+
+### Removed
+
+### Fixed
+
+### Chore
+
+## [1.34.0] - 2019-09-20
+
+### Added
+
+-   Tokei Importer #538
 -   Prominent Notice that we use Sonar-jar #713
 
 ### Changed
@@ -19,6 +33,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Fixed
 
 ### Chore
+
+-   Bump kotlin-reflect from 1.3.41 to 1.3.50 in /analysis
+-   Bump json from 20180813 to 20190722 in /analysis
+-   Bump rxjava from 2.2.9 to 2.2.12 in /analysis
+-   Bump assertj-core from 3.12.2 to 3.13.2 in /analysis
+-   Bump sonar-java-plugin from 5.12.1.17771 to 5.14.0.18788 in /analysis
 
 ## [1.33.0] - 2019-09-10